Added bp-loader.php to the repo, that will selectively load components based on what the user has enabled/disabled via the "BuddyPress > Component Setup" admin panel.
BuddyPress is now just one plugin to activate, once activated you can disable components via the menu mentioned above. This has been done to provide better support for the auto-update functionality, and to try to reduce the "plugin suite" confusion. Each component is still a separate plugin, but they are just not activated individually.

Also added a "BuddyPress" top level navigation item to the admin panel. BuddyPress settings are all now contained under this menu. "Groups" still sits in site admin though, as it fits in well with the "Users" and "Blogs" menus.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@1311 cdf35c40-ae34-48e0-9cc9-0c9da1808c22