In bp_get_the_profile_field_options(), don't delete properties from $field global
bp_get_the_profile_field_options() expects the $field global object to be an
instance of BP_XProfile_Field, because it needs access to that class's
get_children() method. Inside of a profile loop, however, $field is *not* an
instance of this class. So bp_get_the_profile_field_options() clears $field and
replaces it with a new instance of BP_XProfile_Field. However, this has the
side effect of wiping out any custom data that has been manually inserted into
the $field global (for instance, data and visibility_level). This meant that
visibility level was being reset in the case of radio button and other
get_children() fields.

This changeset works around the problem by ensuring that the new
BP_XProfile_Field object contains all of the properties of the $field global
before it is replaced; when they are missing, they are supplied by the
manually-created $field global before it is replaced with the new object.

It will be a good idea down the road to revisit the way that the profile loop
is set up, so that the BP_XProfile_Field is used as much as possible, making
this kind of switcheroo unnecessary.

See #4233

git-svn-id: http://buddypress.svn.wordpress.org/trunk@6069 cdf35c40-ae34-48e0-9cc9-0c9da1808c22