In xprofile_admin_manage_field(), make sure $field_id is set before setting meta
When creating a new profile field, it's necessary to ensure that $field_id is
set, or calls to bp_xprofile_update_meta() and related wrapper functions will
fail. That means that, eg, field visibility settings are not properly stored
on the initial field creation. This changeset fixes the problem by ensuring
that $field_id has been populated before carrying on with the setup routine.

See #4275

Props imath

git-svn-id: http://buddypress.svn.wordpress.org/trunk@6117 cdf35c40-ae34-48e0-9cc9-0c9da1808c22