Do a strict type check for 'html' parameter in bp_core_fetch_avatar()
In r6046, the type check was loosened, to make it somewhat easier for themers
to pass a false value for 'html'. However, it was found after the 1.6 release
that this change broke backward compatibility with certain ways of calling
bp_core_fetch_avatar(). See #4215 for info on the original change.

Fixes #6046

Props rogercoathup

git-svn-id: http://buddypress.svn.wordpress.org/trunk@6247 cdf35c40-ae34-48e0-9cc9-0c9da1808c22