Improvements to automatic WP page mapping for BP directory components
Since the wizard was removed, and Members and Activity were set as the default
components, BP has been creating directory pages for Members and Activity
automatically upon installation. This created a jarring disparity for the
Groups component, which also needs a top-level directory; when activated at a
later date, the user would be prompted to associate a page manually.

This changeset moves bp_updater_add_page_mappings() to the core component
(renaming it appropriately). This function is refactored to allow for existing
page mappings to be retained, as you'd want to do when activating new
components (rather than deleting existing mappings and mapped pages, as you'd
want during a fresh installation). Finally, the function is called when saving
the Components admin page, so that components requiring directory pages (like
Groups) get them created automatically upon activation.

Fixes #4805

git-svn-id: http://buddypress.svn.wordpress.org/trunk@6770 cdf35c40-ae34-48e0-9cc9-0c9da1808c22