Improves the save routine for the Components Dashboard page
The subtabs of the Components Dashboard page (All, Active, Inactive, Must Use,
Retired) all work differently in terms of how data is returned from them. Some
of the subtabs - All and Active - are accurate reflections of the components
that the user currently wants to be active. Retired and Inactive, on the other
hand, don't contain complete data about which components are supposed to be
turned on. For that reason, we can't trust the checkbox data submitted through
the POST when setting the active components.

This changeset introduces bp_core_admin_get_active_components_from_submitted_settings(),
which handles the logic necessary to determine which components the user
intends to be active, based on the data submitted in the POST request. It's
been separated out into a standalone function for better unit testing.

Test: https://github.com/buddypress/BuddyPress-Unit-Tests/commit/49e72ea3fc1f103ab6336c79f5c5b66294f4aad6

See #4671

git-svn-id: http://buddypress.svn.wordpress.org/trunk@6842 cdf35c40-ae34-48e0-9cc9-0c9da1808c22