Moves automated testing suite into BuddyPress core
The new, improved BuddyPress testing suite boasts the following features:
- It uses a vanilla copy of the WordPress test suite as a basis, for maximum
  portability and future-compatibility
- Allows for BP-dependent plugins to write their own tests that use the BP test
  libraries (like our data factories)
- Conforms to current best practices for building WordPress plugin tests

The existing testcases themselves will be ported over in the next changeset, so
as to keep this one reasonably small.

See the Codex page (forthcoming) on how to set up and run the tests, how to add
to them, and how to write BP-dependent tests in your own plugin.

See #4889

Props djpaul, boonebgorges

git-svn-id: http://buddypress.svn.wordpress.org/trunk@6905 cdf35c40-ae34-48e0-9cc9-0c9da1808c22