Only pass BP-related cookies to BP AJAX queries.
Previously, we would pass all cookies to admin-ajax.php.  This could
potentially lead to a '414 Long Request' error because of the length of the
querystring, resulting in the AJAX request failing.

This commit only passes cookies with the 'bp-' prefix to admin-ajax.php,
thus limiting the length of the URI request.

Fixes #4606.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@6940 cdf35c40-ae34-48e0-9cc9-0c9da1808c22