Introduces bp_get_send_message_button_args filter
Most buttons have a filter on the args that are passed to bp_has_button().
The bp-messages Send Message button filter 'bp_get_send_message_button', in
contrast, is on the HTML return value of bp_has_button(). This was probably
a mistake in the original implementation. However, simply moving the existing
filter would break compatibility with existing implementations that use it. So
we've added a new filter, which works the same as other button filters, but has
a slightly different naming convention.

Fixes #4536

Props megainfo, rogercoathup

git-svn-id: http://buddypress.svn.wordpress.org/trunk@6943 cdf35c40-ae34-48e0-9cc9-0c9da1808c22