When testing methods that return more than one user, set distinct last_active for each
The default last_activity value for a user created in our factory is equal to
bp_core_current_time(). When creating multiple users in a test, it's usually
the case that all users will be created within the same clock second, so that
their last_activity value is the same. As a result, it's usually the case that
'active' queries will return users in the order they were created in, as a
fallback. However, in some cases (as on travis-ci), the tests run a bit more
slowly, so that later users have later last_activity times, which means that
return arrays are reordered somewhat. By manually setting distinct last_active
values for each created user, we can be precise when setting the order of our
expected user id arrays.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7033 cdf35c40-ae34-48e0-9cc9-0c9da1808c22