In test_bp_has_members_friendship_requests(), make current user super admin before redirecting
This test involves rerouting to a user's friendship request page, which is only
visible to that user and to bp_moderate users. This set of tests is run as an
'administrator' user, which does not have the bp_moderate cap on multisite
installations, so we need to grant super admin priviliges to avoid failures
when attempting to load the user's requests page.

In theory, we should be able to set the current user to the owner of the
requests page. I ran into problems doing this, which appear to be related to
the way that our test suite's go_to() method tears down and sets up current
user data. This should be fixed for real in the future. This changeset is a
workaround until then.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7045 cdf35c40-ae34-48e0-9cc9-0c9da1808c22