Introduces automatic download of BuddyPress translations from translate.wordpress.org. Fixes #4857
BuddyPress now uses cron to periodically check if the version of the current locale's translation on translate.wordpress.org is more recent than the installed version.
If an update is available, the "Updates" badge's count in the WordPress admin menu is incremented, and a new button on the Updates screen allows a site admin to update the translation to that latest version.
This improves user experience for everyone who wants to use BuddyPress in their native language.

Thanks to markoheijnen for help making changes to GlotPress to enable this functionality.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7097 cdf35c40-ae34-48e0-9cc9-0c9da1808c22