Removes bbp_ copypasta in bp-core-options.php
This erroneous code appears to have been in place for several versions. It's
likely that, as a result, the options in question have been broken for several
versions, since the setter and checker were using different option names. So
there should be no backward compatibility concerns with making these changes.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7139 cdf35c40-ae34-48e0-9cc9-0c9da1808c22