Use bp_moderate cap for Activity and Groups panels
When adding the Activity and Groups administration panels using
add_menu_page(), use the 'bp_moderate' cap rather than 'manage_options'
(which was itself a stand-in for our old is_super_admin() checks).
This allows access to the panels to be granted to non-super-admin users
via plugin. The change also means that we can remove the redundant
current_user_can() check at the beginning of the methods in question.

Fixes #5115

Props ericlewis

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7312 cdf35c40-ae34-48e0-9cc9-0c9da1808c22