Add checks for member and activity components on the welcome page.
On certain setups, the default components (members and activity) can be
deactivated programatically.  Since we were not explicitly checking if
these components were active or not, a fatal error occurred.

This commit adds checks for the default components and adds a fallback
message if these components are not active on the welcome page.

Fixes #5181.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7455 cdf35c40-ae34-48e0-9cc9-0c9da1808c22