Fix trailing 'members' slug when non-admins view a spammer's profile.
If we're on a spammer's profile page, only users with the 'bp_moderate'
cap can view subpages on the spammer's profile.

Users without the cap trying to access a spammer's subnav page should be
redirected to the root of the spammer's profile page.

This commit does this by removing the component in the canonical stack
when a user without the 'bp_moderate' cap is on a spammer's profile page.

Fixes #5207.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7456 cdf35c40-ae34-48e0-9cc9-0c9da1808c22