Use 'user_nicename' when adding a member via autocomplete in the group admin dashboard page.
Previously, 'user_login' was used, which is an unsanitized version of the
username.  When attempting to add a member with special characters (such
as an email address), the JS would throw an error.

Switching to 'user_nicename' allows a sanitized version of the username to
be used and fixes the issue with the autocomplete JS.

Fixes #5187.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7478 cdf35c40-ae34-48e0-9cc9-0c9da1808c22