Ensure that xprofile fields have a default can_delete value of 1.
[7419] fixed a long-standing bug that prevented the can_delete value from being
saved when creating/editing an xprofile field. However, it uncovered a new bug,
which is that new xprofile fields were being created without specifying that
the default value of `can_delete` should be 1. As a result, all new fields
were being created as non-deletable (and, in fact, non-editable; see #5237 for
more details).

See #5237

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7563 cdf35c40-ae34-48e0-9cc9-0c9da1808c22