Use the correct root_id when composing the activity Admin screen
Activity comments have the root_id of the root parent item, while non-comments
have their own ID as the root_id. Setting this value correctly ensures that the
correct data is sent with AJAX comment creation, which in turn ensures that
comments are associated with the correct activity item.

Fixes #5014

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7579 cdf35c40-ae34-48e0-9cc9-0c9da1808c22