In test suite, break constant definitions into separate file
BP's test installation routine is generally run from bootstrap.php. But in
certain cases, as when the suite is being invoked from a BP-dependent plugin's
tests, loader.php can be called directly, skipping bootstrap.php. For this
reason, it makes sense to define constants in a separate file and then include
that file as needed when the tests are first loaded, via either path.


git-svn-id: http://buddypress.svn.wordpress.org/trunk@7665 cdf35c40-ae34-48e0-9cc9-0c9da1808c22