Sanity check before deleting old avatar files before cropping new ones
Previously, if a user uploaded a new avatar file that had a name format like
an existing BP avatar file, BP would accidentally wipe that file out before
cropping. This would result in a failed crop, because the just-uploaded image
would be gone. This changeset introduces a sanity check: only delete existing
avatar files if the filename for those files is not the same as the uploaded
file.

Fixes #5417

git-svn-id: http://buddypress.svn.wordpress.org/trunk@7943 cdf35c40-ae34-48e0-9cc9-0c9da1808c22