Unit Tests: on travis-ci with PHP 5.5.*, don't run the tests with `WP_DEBUG=true`.
PHP 5.5+ uses the `ext/mysqli` library and as versions of WordPress older than the current trunk don't support `mysqli`, PHP throws a bunch of `E_DEPRECATED` warnings about the use of the deprecated `ext/mysql` library. Because we test several versions of WordPress against several versions of PHP, combinations occur with old WordPress and new PHP, which causes these deprecation notices, causing failures.

A better solution would be to somehow suppress `E_DEPRECATED` warnings in the phpUnit configuration or change the error levels that `WP_DEBUG` sets, but so far I've been unable to find a way to do this. For now, travis-ci tests against PHP 5.5 will run with `WP_DEBUG=false`; this is actually the environment we've been testing in on travis-ci until very recently (see r8015).

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8018 cdf35c40-ae34-48e0-9cc9-0c9da1808c22