Set bp_notifications_add_notification() arguments correctly in case of a new_membership_request
The group id needs to be the item_id argument of bp_notifications_add_notification() in case a member requested to join the group. The patch also makes sure the link to the user notification settings page is the one of the administrator of the group in the email sent to inform him of a new membership request.

Fixes #5446

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8037 cdf35c40-ae34-48e0-9cc9-0c9da1808c22