Theme compatibility: Fix issues with themes using post type conditionals.
Previously, when BP resets a post for theme compatibility, we set the
'post_type' to a post type that does not exist (eg. 'bp_activity').  For
themes doing conditional template loading based on the post type, BP's
theme compatibility would not kick in due to the non-existent post type.

This commit replaces all 'post_type' values in bp_theme_compat_reset_post()
to 'page' to address this problem.

For backward compatibility, this commit also modifies the post class to
emulate the older post classes prior to this commit resulting from the
'page' post type change.

Fixes #5418.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8682 cdf35c40-ae34-48e0-9cc9-0c9da1808c22