Core: adjust the SQL generated by `BP_User_Query`'s `search_wildcard` parameter for better Suggestions API search results.
Introduced in r8675, the `search_wildcard` parameter controls where BuddyPress places MySQL wildcard characters around the search term (when it's set).

This change adjusts how the wildcard SQL is constructed; in addition to the wildcard, we now also check for matches that have a single space character at the opposite end of the string (from where the wildcard is). This allows searching data that looks like (for example) a surname in a combined first name/surname field.
e.g. if a displayname is "Paul Gibbs", you can search for "Gi", and with `search_wildcard=right`, the SQL generated will be `LIKE 'Gi%' OR LIKE '% Gi%'`.

See #3278

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8753 cdf35c40-ae34-48e0-9cc9-0c9da1808c22