Fix `bp_is_current_component_core()` to reference BP core components.
`bp_is_current_component_core()` is a function that is meant to check
whether the current component is a BP bundled core component.  This is done
by grabbing `bp_get_option( 'bp-active-components' )`, which is supposed to
hold only BP core components.  However, if a plugin manually saves its
component in the `'bp-active-components'` DB option, our function will
now check this 3rd-party component as well, which is incorrect.

To address this issue, this commit introduces `bp_core_get_components()`,
which is a hardcoded array of BP core components, and references this
function in `bp_is_current_component_core()`.  This ensures that only
BP core components are checked.

See #5552.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8830 cdf35c40-ae34-48e0-9cc9-0c9da1808c22