Messages: don't make links in message excerpts clickable.
When a user is on the inbox screen viewing the list of inbox messages, clicking on the excerpt usually takes them to the message thread conversation page where they can read the full message content. If the excerpt part of the full message content happens to have a link in it, the user could end up on some external website after clicking the excerpt link (which 99% of the time takes them to the full message content).

This change removes `make_clickable` from operating on the message excerpt, fixing this problem.

Fixes #5368, props tw2113

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8921 cdf35c40-ae34-48e0-9cc9-0c9da1808c22