Make sure the updated activity after a post update is consistent with the initial activity.
In Blogs component, when recording a new_blog_post" update, {{{bp_activity_thumbnail_content_images()}}} and {{{bp_create_excerpt()}}} are used to build the content of the activity. After a updating the post from the dashboard, {{{bp_blogs_update_post()}}} was simply updating the activity with the entire post content.

{{{bp_blogs_update_post()}}} is now using the same functions than {{{bp_blogs_record_activity()}}} to update the content of the activity.

Fixes #5792

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8929 cdf35c40-ae34-48e0-9cc9-0c9da1808c22