Do not use `get_comments()` with `fields=ids` parameter.
`fields=ids` is only supported in WordPress 4.0:
https://core.trac.wordpress.org/ticket/28434

This was causing unit test failures in older versions of WordPress.

To workaround this, this commit introduces `bp_blogs_comments_clauses_select_by_id()`,
which is used as a filter so we can select comments by the `comment_ID`
column only and is an alternative to the `fields=ids` approach.

Fixes #5609.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@8945 cdf35c40-ae34-48e0-9cc9-0c9da1808c22