Activity: @mentions Fix search logic for RTL.
I'd previously made a faulty assumption regarding letter order in the database, but my testing reveals the letter order stays intact, it is the word order that is reversed. Pending feedback from anyone who uses a RTL language, I'm modifying the search logic so that it's identical in both LTR and RTL modes.

Fixes #5855

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9009 cdf35c40-ae34-48e0-9cc9-0c9da1808c22