Better detection for false positives in meta SQL filters.
Our wrappers for WP's _metadata() functions require some filtering of the SQL
string (to change a column name and, in the case of xprofile, to add an
'object_type' clause). Our str_replace() logic is too generous, creating the
possibility of matching quoted text, as when the meta value contains the string
'WHERE'.

This changeset modifies the filters so that quoted content is swapped out with
placeholders before we run our search-and-replace.

See #5919.
Props tometzky for feedback.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9073 cdf35c40-ae34-48e0-9cc9-0c9da1808c22