Pass `$pag_sql` variable to 'bp_activity_get_user_join_filter'.
There are two instances of this filter, and the first one had this `$pag_sql`
variable passed to it, while it was missing in the second case (where there is
no pagination). For greater consistency in documentation and plugin
development, we match the format of the variables passed to each instance of
the filter.

See #5938.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9087 cdf35c40-ae34-48e0-9cc9-0c9da1808c22