Core: partial revert of r9116 (user->id to user->ID changes).
In our legacy user fetching functions, we use something similar to "select ... ID as id" in the SQL. In the reverted changes, we are only using the property to return an array of integers, and the change of the property name was causing PHP to throw "undefined property" messages.

As these lines would have never thrown a WordPress core deprecated warning about using `id` instead of `ID`, I'm changing them back to clear the warning message and fix the unit tests.

See #5979

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9119 cdf35c40-ae34-48e0-9cc9-0c9da1808c22