Remove all links in titles of directory and single item pages
Since BuddyPress 1.7, to display the "Create" buttons (Blogs & Groups component) and links to the single item landing pages at a similar place to BP-Default, BP Theme Compat was adding links into the title of the WordPress post object. For 2.2, team has decided to remove all these links and find a better placement for them before the end of the dev-cycle.

See #6008

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9157 cdf35c40-ae34-48e0-9cc9-0c9da1808c22