When cropping an avatar, make sure the "w" crop parameter does not interfere with WP_Query
In the different edit-avatar templates, cropping forms include a field named "w". When posted, this particular posted variable is interpreted as the week parameter of the main WordPress query. This commit introduces a new function bp_core_avatar_reset_query() that will make sure the week parameter is not set when an avatar is being cropped.

props boonebgorges, DJPaul

Fixes #5999

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9158 cdf35c40-ae34-48e0-9cc9-0c9da1808c22