XProfile: Make sure `xprofile_get_field_id_from_name()` returns a field ID for a field name.
Previously, when using `xprofile_get_field_id_from_name()`, it was possible
to get the option ID instead of the desired field ID because a field name
and an option value (for a checkbox or radio field) can potentially be the
the same in the `wp_bp_xprofile_fields` DB table.

This commit ensures that field name look ups will return a field ID instead
of an option ID.  Commit also includes a unit test.

Props dontdream.

Fixes #6082.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9324 cdf35c40-ae34-48e0-9cc9-0c9da1808c22