Search consistently against blog title and description in `BP_Blogs_Blog::get()`.
Previously, the description field was only being matched for the 'total' query,
with the result that pagination and other count-based results did not match the
blogs actually returned by the query.

This changeset reworks the syntax a bit, to avoid the duplicated logic when
using search terms, and to use proper JOIN syntax instead of commas.

Props jreeve.
Fixes #5858.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9328 cdf35c40-ae34-48e0-9cc9-0c9da1808c22