In `bp_activity_generate_action_string()`, use `bp_activity_get_actions()` rather than touching global directly.
This change ensures that the filters called by `bp_activity_get_actions()` are
run, which gives a chance for dynamically registered actions to register
themselves just in time whenever an activity string needs to be generated.

This change is a more general fix than the one introduced in [9366], so that
changeset (minus its unit tests) has been rolled back. See #6126.

Fixes #6141.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9382 cdf35c40-ae34-48e0-9cc9-0c9da1808c22