Build: add SCSS linter (`grunt-scss-lint`) to Grunt build process.
As part of the companion stylesheet changes for BP 2.3 (#6291 #6338), we'll be adding SCSS files to `src/bp-templates/bp-legacy/css/` to help modernise our CSS workflow. This change adds a SCSS linter to help us maintain our code quality.

An empty `src/bp-templates/bp-legacy/css/twentyfifteen.scss` file has been added because the linter hangs if you tell it to run in a folder that does not contain any SCSS files. This will be replaced shortly with the real companion stylesheet for `twentyfifteen` (#6291).

As `grunt-scss-lint` relies on a ruby gem, the Travis-CI config has been modified to install it before running tests.

Fixes #6351, props netweb.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9697 cdf35c40-ae34-48e0-9cc9-0c9da1808c22