Avatar UI: clean up the Camera Backbone template
The template was using a "hide" class for the save button view with no corresponding css rule. As the strategy is to display the button and eventually inform the user to first click on the "capture" button if the "save" button was first clicked, we are removing this class.

Props hnla for the good catch.

See #6290

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9785 cdf35c40-ae34-48e0-9cc9-0c9da1808c22