Set `svn-autoprops` and `svn:global-ignores`.
These commands were introduced with SVN 1.8 and let us set common properties for every file (and future file) in the project. This avoids setting the same properties on most of our files, and also saves us time by not having to remember to set the properties as new files are added to BuddyPress.

See r9819 where the old properties were cleared. Fixes #6409.


git-svn-id: http://buddypress.svn.wordpress.org/trunk@9820 cdf35c40-ae34-48e0-9cc9-0c9da1808c22