Core: Introduce ability for components to register new features.
This commit allows plugin developers to register a feature in the
`'BP_Component'` class.  To BuddyPress, a feature is merely an internal
marker.  It is up to the developer to implement the feature.

Developers can check if a feature is registered with this snippet:

     `bp_is_active( $component, $feature_name )`

See #6331 where this idea came about.

Props boonebgorges, imath, r-a-y.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@9844 cdf35c40-ae34-48e0-9cc9-0c9da1808c22