Core: Fix issue with `bp_verify_nonce_request()` when a port is used in the home URL.
When a site is using a port for their home URL, `bp_verify_nonce_request()`
can sometimes fail.  This is because the port is duplicated twice due to
`$_SERVER['HTTP_HOST']` including the port number and we're appending the
port number again with `$_SERVER['SERVER_PORT']`.

This commit appends `$_SERVER['SERVER_PORT']` if `$_SERVER['HTTP_HOST']`
does not already include the port number and includes a unit test.

Fixes #6528.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10044 cdf35c40-ae34-48e0-9cc9-0c9da1808c22