Remove duplicate IDs from xProfiles admin screen.
Duplicate `id` values are generated in both the `fieldset` and in any of the 
associated form control elements in the Profile Fields admin screen. The `id` 
value  is required for all form controls to be associated with `label` tags  
while the ‘id` value for the `fieldset` element on the other hand is not 
used in either the Profile Fields admin CSS nor JS. Removing the `id` attribute 
from the `fieldset` element fixes the issue of duplicate `id`s in the DOM .
 
Props mercime, boonebgorges, djpaul, johnjamesjacoby.
Fixes #6472.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10071 cdf35c40-ae34-48e0-9cc9-0c9da1808c22