Lazy-load visibility properties in `BP_XProfile_Field`.
Previously, populating a new `BP_XProfile_Field` object would result in two
queries: one to fetch the field row from the database, and one to populate the
field meta cache in order to set the 'allow_custom_visibility' and
'default_visibility' properties. It's not often the case that we need the
latter properties, which means that the cache-priming query is wasted.

This changeset addresses the shortcoming by loading the properties only when
they're requested:

* `$allow_custom_visibility` and `$default_visibility` are now protected properties of the object. This directs requests for them to the magic methods.
* Introduce `__isset()` and `__get()` to handle backward compatibility.
* Introduce public getter functions for each property.

See #6638.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10164 cdf35c40-ae34-48e0-9cc9-0c9da1808c22