xprofile: remove is_admin() logic from required field check
The function should return the same thing for everyone, as it’s a
property of the field only (not the field-user combo). In the relevant
places where these functions are used in BP - namely, during form
submission - bp_moderate users are able to bypass the requirement.

This change breaks backwards compatibility in cases where the
field_is_required() functions are called directly, and the plugin
expects the function to return true for bp_moderate users. This is
unlikely, and given that if there were, they’d run into this same bug,
let’s fix it.

Fixes #6520

Props boonebgorges

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10232 cdf35c40-ae34-48e0-9cc9-0c9da1808c22