Blogs: don't duplicate activity comments during post/activity sync
When JS is disabled, when you post an activity comment to a blog post
activity item, two activity comments will be created. This happens
because we are updating the activity item three times, and updating the
post comment twice.

This change unhooks some of the actions during activity->blog sync to
prevent these unnecessary updates and the duplicated activity comment.

Fixes #6494

Props imath, r-a-y

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10252 cdf35c40-ae34-48e0-9cc9-0c9da1808c22