Support for `WP_Hook` in `bp_get_template_stack()`.
The proposed `WP_Hook` #WP17817 breaks backward compatibility for plugins that
attempt to manipulate `$wp_filter` in certain ways. For BuddyPress, it means
that we cannot directly modify the callbacks that are currently registered as
an array at `$wp_filter[ $tag ]`, in `bp_get_template_stack()`. Instead, when
`WP_Hook` is not available, we assign the callbacks to a variable (leveraging
`WP_Hook`'s `ArrayAccess`) and manipulate the variable.

Props jbrinley.
Fixes #6649.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10279 cdf35c40-ae34-48e0-9cc9-0c9da1808c22