Pass group objects to some group template filters.
A number of group template functions accept a `$group` parameter, but are
generally used within the group loop, and infer the current group when no
object is explicitly passed to the function. Filter callbacks can generally
assume which group is intended when the filters are invoked within the loop
(by examining the `$groups_template` global), but this technique doesn't work
when the functions are used outside of the loop, with a `$group` passed as
a function parameter. In each of these cases, we should be passing the `$group`
object - whether inferred or explicit.

Props garrett-eclipse.
Fixes #6752.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10383 cdf35c40-ae34-48e0-9cc9-0c9da1808c22