Improve performance of `bp_sort_by_key()`.
`bp_sort_by_key()` originally used `create_function()` to allow sorting by
dynamic `$type` and `$key`. However, `create_function()` has a tendency to
perform poorly. PHP 5.3 anonymous closures are much faster, but are unavailable
as long as we support PHP 5.2. So we introduce a new class that plays a role
similar to an anonymous closure, with a similar speed benefit.

Fixes #6864.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10514 cdf35c40-ae34-48e0-9cc9-0c9da1808c22