Emails: update template for WP 3.9 support
`wp_footer()` is included in the email template for when the template
has been loaded inside the Customiser preview frame. The function used
to for this, `is_customize_preview`, was only introduced in WP 4.0 and
is causing fatal errors in WP 3.9.

The rest of the Customizer integration code already checks for WP 4.0+;
this part of the template was just overlooked.

Fixes #6874

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10535 cdf35c40-ae34-48e0-9cc9-0c9da1808c22