Emails: add email menus to network admin for network-activated configs.
This provides consistency with the rest of BuddyPress; namely, that when network activated, you manage all BuddyPress content from the network admin.
With post types and taxonomies, we have to do a little extra work to add links to the root site on the network-admin, because of API limitations in WordPress.

See #6592

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10538 cdf35c40-ae34-48e0-9cc9-0c9da1808c22