Messages: Return generic WP Error in `messages_new_message()` when necessary.
In #6535, we added an `'error_type'` parameter in `messages_new_message()`
that supports returning a WP Error object if error type is set to
`'wp_error'`.  However, we did not return a WP error further down in the
`messages_new_message()` function, which can result in a potential fatal
error to occur.

This commit rectifies this and includes a unit test.

Anti-props r-a-y.

See #6894 (trunk).

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10585 cdf35c40-ae34-48e0-9cc9-0c9da1808c22