Emails: set default value of `unsubscribe` token to `wp-login.php`.
If the email recipient can't be found by `get_user_by( 'email', ... )` on multisite configurations or when `BP_SIGNUPS_SKIP_USER_CREATION` is enabled, the unsubscribe link in the emails is empty.

In a future release, we may introduce logic in the template to not render this token if it is valueless, but for 2.5, we're setting it to the log in form. If a user who hasn't activated their account tries to log in, they will see a message informing them they need to activate their account first, which is good enough.

See #6915

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10611 cdf35c40-ae34-48e0-9cc9-0c9da1808c22