Notifications: Introduce two new parameters to the `'bp_notifications_get_notifications_for_user'` filter.
This commit:
* Adds `$component_action_name` and `$component_name` as the 6th and 7th parameters to the filter.
* Updates the DocBlock for the filter.

In the DocBlock, we are deprecating the first parameter, which is the
component action name.  It is recommended to do notification action name
checks against the 6th parameter from now on.

The reason why we are deprecating the first parameter is due to potential
plugin conflicts. If a plugin is doing a specific check on the notification
action and changes the return value for the filter, then no other plugin
can do checks for that same notification action.  Also see #6669.

Fixes #7020.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10725 cdf35c40-ae34-48e0-9cc9-0c9da1808c22