Use absint() to validate numeric input in the WP Customizer.
`intval()` fails when used with the WP
Customizer API as a sanitization callback.
The callback function is attached to
a filter in `WP_Customize_Setting::sanitize()`
that passes two arguments to the
sanitization callback: the `$value` to sanitize
and the instance of `WP_Customize_Setting`.
Since `intval()` can accept an optional
second argument ($base), but isn’t
expecting an object, it fails.
`absint()` works because it is only
expecting the `$value` argument.

Fixes #7041.

Props WeddyWood.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10814 cdf35c40-ae34-48e0-9cc9-0c9da1808c22