Taxonomy: Allow the site used for taxonomy term storage to be filtered.
On some setups, it may be desirable to store the terms from a given BP taxonomy
on a site other than the default. For example, on a multi-network setup, where
users are shared between all networks, it may suit the purposes of the
installation better to have bp_member_type data stored on the primary network
and shared between networks.

The new function `bp_get_taxonomy_term_site_id()` function allows the value to
be filtered, on a taxonomy-specific basis. BP's taxonomy wrapper functions have
been updated to use this new function.

Props rekmla.
Fixes #7077.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10818 cdf35c40-ae34-48e0-9cc9-0c9da1808c22