Activity: Fine-tune activity reply screen notifications when activity stream commenting is enabled.
Previously, if "activity stream commenting on blog and forum posts" is
enabled and a reply was created from the Activity Directory page for a
blog post activity item, clicking on the activity reply notification
would not mark the notification as read.

Also, if a member decides to post a comment from the blog post permalink,
a screen notification would not be added.

This commit fixes these issues and adds a unit test.

Props r-a-y, dcavins, imath.

Fixes #7135.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10895 cdf35c40-ae34-48e0-9cc9-0c9da1808c22