Travis-CI: adds code coverage with https://codecov.io
Adds a new Travis-CI task which runs PHPUnit with code coverage enabled.
The results are piped to codecov.io, which we're experimenting with to track
progress on our code's test coverage as the project grows.

Code coverage generation is very slow. To try to minimise the impact on the
rest of our Travis-CI configuration, the new `travis:codecoverage` task is
set in the `allowed_failures` section, which means Travis-CI should report
test pass/fail without waiting for the code coverage generation.

Another effort to speed this up has Travis-CI running the code coverage task
on PHP7 with `phpdbg`, as opposed to relying on `xdebug`, which is the default.

See #7207 and https://codecov.io/gh/buddypress/BuddyPress

git-svn-id: http://buddypress.svn.wordpress.org/trunk@10997 cdf35c40-ae34-48e0-9cc9-0c9da1808c22