Travis-CI: disable xdebug globally and switch `travis:codecoverage` to use xdebug instead of phpdbg.
In r10997, we set up a code coverage task which used phpdbg to speed up code coverage report generation.
Iteration and further testing has shown the Travis-CI environment will not support our particular project
layout with phpdbg. This commit simplifies the code coverage to use xdebug.

xdebug has also been globally disabled on Travis-CI (the code coverage task explicitly re-enables it)
because we are not relying on it for any parts of our existing tests, and doing so should speed up
running the unit tests.

See #7207

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11001 cdf35c40-ae34-48e0-9cc9-0c9da1808c22