Groups: rename `loop_start` and `loop_end` actions.
This gives the Group Invitations and Group Members loops unique names for the actions called at the start and end of those templating loops. These are analagous to WordPress' Post templating loop with `loop_start` and `loop_end`, which is where the action names were likely copied from originally.

This has caused compatibility problems with plugins that hook to these actions that rightly expect the type of the first passed argument to be of `WP_Query`.

Fixes #7247

Props herbovec for reporting the issue.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11083 cdf35c40-ae34-48e0-9cc9-0c9da1808c22