Activity: Move Akismet loader to `bp-activity-akismet.php`.
Previously, `bp_activity_setup_akismet()` was placed in
`bp-activity-actions.php`.  The actions file isn't a proper place to put
this code.

This commit moves this function to `bp-activity-akismet.php` and also
removes some duplicate conditional code. `BP_Activity_Component` already
checks if Akismet is loaded, so no need to re-duplicate this logic.

This is part of some prep work to conditionally load a component's action
and screen code when necessary.

See #7218.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11102 cdf35c40-ae34-48e0-9cc9-0c9da1808c22