Core: Refactor `BP_Button` class to use an easier syntax.
Previously, to use the `BP_Button` class, a dev would have to define
several parameters in order to render the HTML element for a button:

array(
  'wrapper' => 'div',
  'wrapper_id' => 'my-wrapper',
  'wrapper_class' => 'my-wrapper-class',
  'link_href' => 'hxxp://example.com',
  'link_class' => 'my-link-class',
  'link_id' => 'my-link-id',
  'link_rel' => 'nofollow',
  'link_title' => 'my-link-title'
)

This commit simplifies the syntax to:

array(
  'parent_element' => 'div',
  'parent_attr' => array(
    'id' => 'my-wrapper',
    'class' => 'my-wrapper-class'
  ),
  'button_attr' => array(
    'href' => 'hxxp://example.com',
    'class' => 'my-link-class',
    'id' => 'my-link-id',
    'rel' => 'nofollow',
    'title' => 'my-link-title'
  )
)

The `'parent_attr'` and `'button_attr'` parameters can use any arbitrary
HTML attribute set as the array key.

This commit also means we are deprecating the older parameters listed in
the first example.  However, we still support these parameters via backward
compatibility.

Fixes #7226.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11114 cdf35c40-ae34-48e0-9cc9-0c9da1808c22