Stop locating an empty set of templates.
Bail early in `bp_locate_template()` when there are no files to look for.
This prevents PHP from showing errors when an empty array is provided. Use
case for this is when by hooking into the `bp_get_template_part` filter
you'd like to prevent the locating of a template.

Fixes #7149.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11267 cdf35c40-ae34-48e0-9cc9-0c9da1808c22