Prevent BuddyPress from loading when the minimum PHP version is not met.
When the server is running a version of PHP lower than 5.3.0, an
admin notice is thrown, and the rest of BuddyPress is not loaded.

The bp-loader.php file must remain compatible with WP's minimum PHP 
version, to prevent fatal errors when activating BuddyPress. In order
to reduce the cognitive overhead involved in maintaining PHP compat
for this file, the `BuddyPress` class has been moved into its own
class-buddypress.php file.

Props r-a-y.
See #7277.


git-svn-id: http://buddypress.svn.wordpress.org/trunk@11389 cdf35c40-ae34-48e0-9cc9-0c9da1808c22