Activity: Add parameter to `bp_activity_get_comment_depth()`.
Previously, `bp_activity_get_comment_depth()` only supported looking for an
activity comment's depth during a queried activity comment loop.

This commit allows a developer to pass either a comment object or an
activity ID to the function so the activity comment depth can be fetched
outside the activity comment loop.

This fixes an issue with a notice displaying in the BuddyPress Activity
admin dashboard.

See #7329.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11413 cdf35c40-ae34-48e0-9cc9-0c9da1808c22