Messages: Fix redirection issue when attempting to view a valid private message thread.
Previously, attempting to go to a valid private message thread URL when
logged out would throw a 404.

This commit properly redirects non-authenticated users attempting to view
a private message thread to the login screen.  If a user is authenticated,
but is attempting to view a private message thread without access rights,
we redirect the user back to their own "Messages" page with a notice.

See #7349.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11537 cdf35c40-ae34-48e0-9cc9-0c9da1808c22