Restore default behavior of message link generating functions.
Prior to [11578], functions for generating links in the context of
messages (Mark Read, Delete, etc) were generated relative to the
logged-in user. This made it impossible for an administrator to
manage another user's messages. [11578] changed the default behavior
of these functions to reference the displayed user. But this change
breaks backward compatibility for functions expecting the prior
behavior, including BP's own message notification format functions.

We work around the limitation by introducing `$user_id` parameters
to the functions in question. `$user_id` defaults to the logged-in
user (for backward compatibilty), and the bp-legacy template passes
in the displayed user ID as appropriate.

Fixes #7096.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11644 cdf35c40-ae34-48e0-9cc9-0c9da1808c22