Nouveau: Correct non showing of single comment reply in activity stream
in the activity JS we count replies and if less than one we attempt to ensure that we do show a comment where initially all are hidden, this was actioned by comparing counts and removing the 'hide' actions/classes for the last-child element, this fails if the first or only element.

Commit addresses this by removing the `.last()` function.

Todo: This comments section of scripting has proven problematic & this may not be a final solution, more testing would be useful in general for commenst display. 

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11859 cdf35c40-ae34-48e0-9cc9-0c9da1808c22