Groups: Conditionally load action and screen functions.
This commit conditionally loads action and screen function code for the
Groups component, utilizing the `'bp_late_include'` hook introduced in
r11884.

Previously, we loaded these functions at all times, which is unnecessary
when a user is not on a BuddyPress groups page. Now, we only load this code
when needed.

See #7218.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11923 cdf35c40-ae34-48e0-9cc9-0c9da1808c22