Core: allow theme to override site's template pack choice.
As part of the work for Nouveau, we need a way to let a theme override the site template pack setting.
e.g. new themes could force the "nouveau" template pack.

The syntax is `buddypress-use-foobar`, where "foobar" is the slug name of the template pack.
e.g. `buddypress-use-legacy`, `buddypress-use-nouveau`.

Fixes #7730

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11936 cdf35c40-ae34-48e0-9cc9-0c9da1808c22