Allow pending invite to be cancelled without friendship requirement.
BP Legacy's function `bp_legacy_theme_ajax_invite_user()` was built to stop early if the invitee was not a friend of the inviter. In the case where you extend an invitation to a user, then cancel your friendship, you were prevented from running the `uninvite` task. This patch removes the friendship check from the uninvite task.

Props r-a-y.

Fixes #7726.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@11969 cdf35c40-ae34-48e0-9cc9-0c9da1808c22