BP Nouveau: Activity directory, remove misleading 0 count from tabs.
Into the Activity directory the count spans added to the scope tabs are there to inform about newest activities (HeartBeat) for each scope. Starting the UI with a count set to 0 can be misleading because we could believe this count is about the total number of activities for each scope.

Fixes #7788

git-svn-id: http://buddypress.svn.wordpress.org/trunk@12034 cdf35c40-ae34-48e0-9cc9-0c9da1808c22