BP Nouveau - Messages UI: Improve a11y in message boxes.
Message boxes are listing messages (inbox, starbox, sentbox). Using screen reader softwares or the keyboard, it is now a better experience to navigate in these lists. Pressing the Tab key makes you navigate from the Message Checkbox, to the sender/recipients link and finally to the Message title. Once on a Message title, pressing the enter key activates the message and updates the preview box. Pressing on the enter key another time opens the full conversation.

Huge Props mercime

Fixes #7731

git-svn-id: http://buddypress.svn.wordpress.org/trunk@12073 cdf35c40-ae34-48e0-9cc9-0c9da1808c22