BP Nouveau: Improve Messages UI extensibility
In BP Nouveau, the Messages UI is a Backbone based one. This means current legacy hooks inserted into the JS templates are a bit more different to use for plugin developers as it requires them to use JavaScript to get the data models. If the current AJAX requests still need a way to be extended to fetch extra data such as specific Messages metas, this commit is a first step to help plugin developers to insert content into the Messages UI without changing their habits about using the PHP Messages template global variables to get data about the Messages loop.  This is done by introducing the back compatibility function {{{bp_nouveau_messages_catch_hook_content()}}}. This function is used during the AJAX requests that are fetching messages for a specific thread. As a start, 2 legacy hooks have been moved from the JS Template level to this higher level :
- {{{bp_before_message_content}}}
- {{{bp_after_message_content}}}

If actions are attached to these hooks their outputs will be caught using the buffer and will be "transported" as new properties, respectively {{{beforeContent}}} and {{{afterContent}}}, of the JSON reply that is used by the Messages UI. If these 2 properties are set and populated, their content will be output inside the corresponding JS templates.

Props pareshradadiya

See #7847 (Trunk)

git-svn-id: http://buddypress.svn.wordpress.org/trunk@12121 cdf35c40-ae34-48e0-9cc9-0c9da1808c22