Messages: catch Legacy hooks to customize the Active conversation box.
In order to preserve in the BP Nouveau Messages UI the level of PHP customization that developers are used to get in Legacy using {{{do_action}}} hooks, The content generated for the following hooks will be caught during Ajax requests and added as new properties of their JSON replies:

- {{{bp_before_message_meta}}}
- {{{bp_after_message_meta}}}
- {{{bp_messages_inbox_list_item}}}
- {{{bp_messages_thread_options}}}

About the {{{bp_messages_inbox_list_item}}} hook: as the Messages UI of BP Nouveau, unlike Legacy, is not using a table tag to display the list of messages received/starred/sent. All extra actions about messages into these loop views are done within the Active conversation box (eg: star/unstar). To inject the content of this specific hook, we are adding a table element under the message content preview as the expected output is a td tag.

See #7851 (Trunk)

git-svn-id: http://buddypress.svn.wordpress.org/trunk@12141 cdf35c40-ae34-48e0-9cc9-0c9da1808c22