Travis CI: Don't install Composer packages on PHP < 5.4.
The phpcompatibility-wp library has requirements that need at least PHP 5.4,
and we only use this library when testing against 7.2, so there's no need
to install on earlier version of PHP.

This will likely have to be revisited at some point in the future, when
we use Composer for more integral parts of the build/dev process.

See #7997.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@12282 cdf35c40-ae34-48e0-9cc9-0c9da1808c22