Improve the `bp.apiRequest()` JS function and some PHP code indentation
`bp.apiRequest()` will make sure to return the JSON response object in case of an error. In case this JSON response object is not available a default error object will be used.

When WordPress version is lower than 4.9, the same function will make sure to override the `url` option with the REST API root URL only if this option is not already set. Moreover the Nonce header is now only added for REST API requests made on the current site domain.

The PHP code indentation of `bp-core/bp-core-rest-api.php` file has been improved.

Fixes #8131


git-svn-id: http://buddypress.svn.wordpress.org/trunk@12454 cdf35c40-ae34-48e0-9cc9-0c9da1808c22