Edit the type of the `bp_block_category()` function's second argument
Although this filter should always get a `WP_Post` as a second argument, it appears there is a risk this function is used directly passing a null value as the second argument. See this [https://buddypress.org/support/topic/since-update-to-5-0-buddypress-freezes-features-of-wordpress-backend/ support topic].

To prevent this risk of notice error, we are making sure the second argument is a `WP_Post` in a different way.

See #8152 (Trunk)


git-svn-id: http://buddypress.svn.wordpress.org/trunk@12490 cdf35c40-ae34-48e0-9cc9-0c9da1808c22