Emails: ensure compatibility with PHPMailer 6.0
Since version 5.5, WordPress has deprecated version 5.2 of PHPMailer in favor of its 6.0 version. We are using this library for our BP Emails feature. To ensure compatibility with this new version of WordPress as well as with the previous versions we are supporting (4.8 to 5.4), we need to make sure to use the right version of PHPMailer according the version of WordPress installed.

Props hareesh-pillai, man4toman

See #8322 (trunk)


git-svn-id: http://buddypress.svn.wordpress.org/trunk@12687 cdf35c40-ae34-48e0-9cc9-0c9da1808c22