Admin: use `implode()` the right way into the BP site health panel
The right order of arguments of this function is `$glue`  first and then `$pieces`, not the other way around as it was so far. Moreover as of PHP8 (Scheduled for Nov 26th), `implode( $glue, $pieces )` only accepts args in that order, it no longer supports passing the `$glue` as the second arg.

Props dd32

See #8367 (trunk)


git-svn-id: http://buddypress.svn.wordpress.org/trunk@12732 cdf35c40-ae34-48e0-9cc9-0c9da1808c22