Declare `$get_data = false` when calling `xprofile_get_field()`.
The default value of `$get_data` is `true`, which means that when calling
`xprofile_get_field()` with no explicit `$user_id`, but with a logged-in user,
`BP_XProfile_Field::populate()` will attempt to fetch the user's data. This
is usually neither intended nor needed, and results in unnecessary database
overhead.

See #8378.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@12768 cdf35c40-ae34-48e0-9cc9-0c9da1808c22