BP Nouveau: make sure Ajax searching in single Group works as expected
When the custom front template is not active, the activity template is loaded as the Group's home page. In this case we need to make sure the search form selectors IDs are including the "activity" reference instead of the "home" one. If the Activity component is not active, Group's home is the members template: we also need to make sur the corresponding search form selectors IDs are including the "member" reference instead of the "home" one.

Props corsky, vapvarun

Fixes #8386


git-svn-id: http://buddypress.svn.wordpress.org/trunk@12774 cdf35c40-ae34-48e0-9cc9-0c9da1808c22