XProfile: only allow "style" attributes in richtext fields for capable users
This commit prevents non-capable users from adding style attributes to "span" and "p" elements in their profile fields, which could be used in unintended ways relative to when it was introduced in #5625.

Note that this could be considered a backwards compatibility break. If you are a site owner or developer who relied on this functionality, you will want to use the `xprofile_allowed_tags` filter to re-enable these attributes.

In trunk for 7.0. See #8404.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@12806 cdf35c40-ae34-48e0-9cc9-0c9da1808c22