BP_Invitation: Increment date_modified on key actions.
When an invitation is sent or accepted, the `date_modified` should be updated.
However, if a `date_modified` has been specified in the invitation or request
creation function, the `date_modified` should be set to that, even if notices
are sent (now) as part of the creation process.
Ensure that `BP_Invitation::get_query_clauses()` handles specified `date_modified` data updates.