xProfile: remove the autolink feature support from the URL field type
Autolinking this field value to a members search is not suitable for this type of field as its goal is to direct the user to the link's location. To avoid admin users confusion, let's remove the Autolink metabox and this feature from the URL field type.

Props topher1kenobe

Fixes #8450



git-svn-id: http://buddypress.svn.wordpress.org/trunk@12878 cdf35c40-ae34-48e0-9cc9-0c9da1808c22