BP Email: introduce "unnamed" salutation
The member invites feature's goal is to invite external users to join a community. The BP Email feature uses a named salutation by default. As a result, the salutation of invitation emails was "Hi email@domain.tld,". The user's name is not yet known and using this kind of salutation is odd. The member invites feature's need is to simply use an "unnamed" salutation eg: "Hi,". To satisfy it we are updating the BP Email type schema with a new property `$named_salutation` which defaults to true. If this property is false, then the sent email will from now on use an unnamed salutation.

**NB**: to prevent potential back compatibiliy issues, each kind of salutation has their own filters
- Use `bp_email_get_unnamed_salutation` to edit the unnamed one.
- Use `bp_email_get_salutation` to edit the default one (named).

Props dcavins for coloring outside the lines a bit 😉

See #8139



git-svn-id: http://buddypress.svn.wordpress.org/trunk@12932 cdf35c40-ae34-48e0-9cc9-0c9da1808c22