Add php fallback for dismissing site-wide notices.
Dismissal of sitewide notices has been handled
solely via JavaScript in the Legacy and Nouveau
template packs. This commit adds a new action
URI at `{user_profile}/messages/notices/dismiss`
that will dismiss the currently active notice for
the logged-in user.

We also add a new template function to create the
nonced link, `bp_get_message_notice_dismiss_link()`,
and make use of it in the core template function
`bp_message_get_notices()` and in the Nouveau
template for site-wide notices. In both cases, `button`
elements have been converted to `anchor` elements
linking to the new dismissal URI, so the action can be
completed even if there is a problem with JavaScript.

Fixes #8505.