Add a new function to check if a Widget/Block is active into a sidebar
Some of our Legacy Widgets are being displayed only if active they have been included into a Sidebar Widget. We are using the `is_active_widget()` WordPress function to check whether it is the case or not. This function doesn't fit in the Widget Blocks case as all Blocks are being inserted into the specific `WP_Widget_Block` widget.

As WordPress 5.8 doesn't provide any equivalent (so far?), we are introducing a new function to do this check: `bp_is_widget_block_active()`. It takes 2 optional arguments.
- `$block_name`: the name of the block you need to check (eg: `'bp/members'`).
- `$widget_id_base`: the Base ID of the widget (eg: `'bp_messages_sitewide_notices_widget'`).

You can use it to check a block, a widget or both.

Fixes #8515



git-svn-id: http://buddypress.svn.wordpress.org/trunk@12997 cdf35c40-ae34-48e0-9cc9-0c9da1808c22