Settings: Make sure changing pwd from the General Screen encrypts it
Using a WP_User object to set the password to update was misleading WordPress `wp_update_user()` function. The new password was interpreted as the current password and was not encrypted.

Passing a regular array to the `wp_update_user()` function avoids the risk of messing with WP User object cache.

Props mandro

Fixes #8539 (trunk)



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13032 cdf35c40-ae34-48e0-9cc9-0c9da1808c22