Make the `new_message` notification filter consistent across formats
The 'bp_messages_' . $amount . '_new_message_notification' dynamic filter is including **6** parameters for the `string` format and **7** for the `array` format. This makes it difficult to use for Plugin developers.

To make things more consistent across formats, we are deprecating this filter in favor of a new dynamic filter `'bp_messages_' . $amount . '_new_message_' . $format . '_notification'`. FYI, possible filter names are:

- `'bp_messages_multiple_new_message_string_notification'`
- `'bp_messages_single_new_message_string_notification'`
- `'bp_messages_multiple_new_message_array_notification'`
- `'bp_messages_single_new_message_array_notification'`

Fixes #8546



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13035 cdf35c40-ae34-48e0-9cc9-0c9da1808c22