Make sure Group nav items positioning includes the Members nav
BP Nouveau includes a specific customizer tool to edit the position of the Group nav items and a customizer setting to add a custom front page to the Group. When this front page is disabled, the Members nav was not included by the specific customizer tool. It's now the case.

Fixes #8548



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13094 cdf35c40-ae34-48e0-9cc9-0c9da1808c22