Make sure deleting an activity comment also deletes its children
This commit also takes care of redirecting the user to the activity single view when clicking on the comment button and there is no comment form available for the corresponding activity. This case can happen on a single member's activity stream when the main activity is an activity comment.

Props yesbutmaybeno

Fixes #8415



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13114 cdf35c40-ae34-48e0-9cc9-0c9da1808c22