BP Blocks: localize dynamic Blocks script data into site's footer
These script data were not printed into themes using the Full Site Editing feature, like Twenty Twenty-Two, which was generated a JavaScript error as soon as the user clicked on one of the links to filter the loop's items (friends, members or groups).
Moving the script localization to the site's footer makes sure the JavaScript error disappears and that links are fully functional. This change has also been successfully tested on regular WordPress themes.

See #8474 



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13194 cdf35c40-ae34-48e0-9cc9-0c9da1808c22