Bring back `custom` order to xProfile field options sorting
r13147 introduced a regression about this type of xProfile field options order. To make sure we preserve it, we need to check `$this->order_by` value before sanitizing it otherwise the potential `custom` value is sanitized to `ASC` and the SQL sort part is never set to use the `option_order` db field.

Props oztaser, espellcaste, niftythree

Fixes #8623 (trunk)



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13231 cdf35c40-ae34-48e0-9cc9-0c9da1808c22