List the displayed user groups invites in member's front-end screen
As a site admin can view the displayed user groups invites, listed invites have to be the one of this user and not the ones of the site admin.

Adapt the Group Invites feature so that site admins can accept or reject on behalf of the displayed user the listed invites. These two actions made by an admin will generate a specific BP Email informing the user of it.

Props oztaser, dcavins, espellcaste

Closes https://github.com/buddypress/buddypress/pull/15
Fixes #8675



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13273 cdf35c40-ae34-48e0-9cc9-0c9da1808c22