Be sure the join/leave group button args exist before trying to use it
This is a follow up to [13302]. When the group creator is displaying the corresponding group's pages, the "join/leave group" header's button is not generated. In [13302], we were wrongly trying to access the arguments of this button although they do not exist. In this particular case, we just need to only perform BP Nouveau's button reset when this button is generated: that is to say when someone else than the creator is viewing a group's page.

See #8722



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13310 cdf35c40-ae34-48e0-9cc9-0c9da1808c22