Avoid querying for comments when viewing a BuddyPress page in TT3
The Twenty Twenty-Three block theme is using the `core/comments` block to 
display comments for WP Pages or Posts. After checking if comments are 
open, this block is doing an additional comment count query which wrongly 
returns the site's all comment count as BuddyPress pages ID is reset to 
`0`.

In this case forcing comments to be closed is not enough, we also need to 
force the comment count to be 0.

See #8744 (trunk)

--Cette ligne, et les suivantes 
ci-dessous, seront ignorées--

M    src/bp-core/bp-core-filters.php
M    src/bp-core/bp-core-theme-compatibility.php


git-svn-id: http://buddypress.svn.wordpress.org/trunk@13332 cdf35c40-ae34-48e0-9cc9-0c9da1808c22