SIte invitations: Do not assume that a user exists.
In some BuddyPress setups, a user is not created
until the signup has been activated. This is the case
when `BP_SIGNUPS_SKIP_USER_CREATION` is set
to `true` and when installed on a WodPress network.
So we rely on the `BP_Signup`  instead of the dodgy
user in `bp_members_invitations_complete_signup()`.