Site invitations: Do not assume that a user exists.
In some BuddyPress setups, a user is not created
until the signup has been activated. This is the case
when `BP_SIGNUPS_SKIP_USER_CREATION` is set
to `true` or when BP is installed on a WodPress network.
So we rely on the `BP_Signup` instead of the dodgy
user_id in `bp_members_invitations_complete_signup()`.

Fixes #8693.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@13346 cdf35c40-ae34-48e0-9cc9-0c9da1808c22