BP Nouveau: Make sure the thread ID is transported into thread view
The Backbone based Messages UI of the Nouveau template pack is now looking for the displayed thread ID inside the messages collection if the current thread object hasn't been returned by the Ajax request. This prevent an error that was happening when trying to reply to a thread message after the thread has been displayed without re-requesting the server.

Props casper99

Closes https://github.com/buddypress/buddypress/pull/63
See #8825



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13415 cdf35c40-ae34-48e0-9cc9-0c9da1808c22