BP Nouveau's message UI: make sure the reply action gets needed data
Adding a reply can be done by opening the thread view from the message UI or by directly accessing to this view from its URL.
In both cases, we need to make sure to get the current Thread ID in order to attach the reply to it.
As the Backbone Model or the fallback object are both including an `id` property, using it is a better choice than trying to access the Backbone Model's `id` attribute.

Props dave.pullig

Closes https://github.com/buddypress/buddypress/pull/89
See #8870 (trunk)



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13457 cdf35c40-ae34-48e0-9cc9-0c9da1808c22