Improve how the `BP_Akismet()` class displays an activity history
When the Akismet plugin is active and is used to prevent BP Activity spams, each Activity item gets an Akismet history. This history is only generated for activities being posted once Akismet is on. As a result, there can be old activities with no history. To avoid potential PHP errors/notices, we need to verify the Activity Meta used to store the history is an array and has a `time` key.

As we are dealing with the latest history entry only, we do not need to sort the Activity Meta.

To display the activity history into its single WP Admin screen, we now need to use the `time` & `message` keys of the Activity Meta.

Props mystichrome

See #8873 (trunk)
Closes https://github.com/buddypress/buddypress/pull/93



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13466 cdf35c40-ae34-48e0-9cc9-0c9da1808c22