Check the requested Member’s page matches a valid nav screen function
The `BP_Members_Component::check_parsed_query()` method happens once the query is parsed & once the single displayed Member's navigation is set. It checks the current action matches a navigation slug & whether the corresponding screen function exists and is callable. If it's not the case a 404 is displayed, just like it's currently the case when the BP Legacy URL parser is in use.

See #4954
Fixes #8932
Closes https://github.com/buddypress/buddypress/pull/125



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13512 cdf35c40-ae34-48e0-9cc9-0c9da1808c22