List all active components in `buddypress()->active_components`
So far this global was only retrieving the list of built-in components that were activated from the BP Components settings screen of the WordPress Dashboard. Third party plugins/add-ons to be in this list needed to add their own component ID to it using this kind of code `buddypress()->active_components['component_id'] = 1`.

BuddyPress is now making sure any active components are listed into the `buddypress()->active_components`.

Fixes #8933
Closes https://github.com/buddypress/buddypress/pull/126



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13513 cdf35c40-ae34-48e0-9cc9-0c9da1808c22