Introduce basic community visibility feature.
Site admins now have the option to restrict access to the
community portion of a BuddyPress site. If the site admin
chooses "members only," then, when a non-logged-in
user attempts to visit a BuddyPress screen, she will
be shown an access error message and a log-in form.

Props imath, sbrajesh, jjj, shawfactor.

Fixes #8734.

git-svn-id: http://buddypress.svn.wordpress.org/trunk@13533 cdf35c40-ae34-48e0-9cc9-0c9da1808c22