Restore the Member’s activity single view
A regression introduced in [13512] was preventing single activities to successfully load into the BP Nouveau Template Pack. We need to only perform the parse query check when viewing a single user. In the case of a single activity there is no need to check for an existing Member’s (sub)navigation item.

See #8932
Fixes #8956
Closes https://github.com/buddypress/buddypress/pull/143



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13534 cdf35c40-ae34-48e0-9cc9-0c9da1808c22