xProfile Admin: use `textbox` as the default selected field type
In [13414] the `BP_XProfile_Field::$type` variable default value was set to `''` to prevent PHP deprecation notices. This move generated a side effect on the selected default field of the select box used to define a new xProfile field type. 

Instead of having the `textbox` field type selected by default, the `checkbox` was used. The downside of this is: this field type, unlike the `textbox` one, needs a meta box to define option values which was not displayed. This meta box is only displayed using JavaScript when the selected value of the Field type select box has changed.

Props strategio

See #9035 (trunk)
Closes https://github.com/buddypress/buddypress/pull/200



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13668 cdf35c40-ae34-48e0-9cc9-0c9da1808c22