BP Member’s nav: reset primary nav item links when a user is displayed
Since 12.0, the member’s nav item links are no more initialized with the logged-in user links but are left empty and built using the BP Rewrites API. It generated a compatibility issue when the BP Legacy URL parser is in use (eg: when the BP Classic Add-on is active): the BP Member’s nav used to populate a custom WP Nav Menu was wrongly using the displayed member’s links instead of the logged-in user ones.

Making sure to reset these links in favor of the logged-in user ones in such a case is now necessary during the custom WP Nav Menu generation process.

Props vapvarun espellcaste

See #9108 (trunk)
Closes https://github.com/buddypress/buddypress/pull/237



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13744 cdf35c40-ae34-48e0-9cc9-0c9da1808c22