Groups: joining a group needs to wait for group status to be defined
When the group create first step is completed, the group's status has not been defined by the group creator yet and defaults to public. As the group status & the invite status are set once the group create second step is completed, we need to wait for this step to be achieved to let users join the group being created otherwise it would be possible for a user to "pre-join" a private/hidden group. Checking if the invite status is set is the only way to make sure this second step has been completed. If it's not the case, no need to go further in `groups_join_group()`.

Fixes #9162
Closes https://github.com/buddypress/buddypress/pull/294



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13874 cdf35c40-ae34-48e0-9cc9-0c9da1808c22