Administration: add a function to check a BP Admin Screen is displayed
`bp_is_admin()` accepts an optional parameter: a string containing the BP Admin Screen ID. If provided: it checks this screen is loaded, if not it only checks a BP Admin Screen is displayed. This function will help us put together a specific community dashboard in the future.

For now it's helping us making sure the WP Plugin dependency warning is not displayed into the WP Admin > Add new plugin > BP Add-ons tab as this tab is generated by BuddyPress: so it is available & activated!

Props emaralive.

Fixes #9212
Closes https://github.com/buddypress/buddypress/pull/334



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13974 cdf35c40-ae34-48e0-9cc9-0c9da1808c22