Enforce requirement of usernames having valid letters.
Profile links do not work for users with usernames that do not contain any valid letters. We have logic in place to check for letters but it is faulty. This commit enforces the requirement that usernames must contain a minimum of 4 and a maximum of 60 characters, with valid letters.

It comes with unit tests and includes some minor coding styles fixes.

Commit to correct the one previously sent at [13984]

Props niftythree.

Fixes #8781
Closes https://github.com/buddypress/buddypress/pull/332/

git-svn-id: http://buddypress.svn.wordpress.org/trunk@13985 cdf35c40-ae34-48e0-9cc9-0c9da1808c22