Make sure `12.0` deprecated code is loaded until version 15.0
Version 12.0 changed BuddyPress a lot with the introduction of the BP Rewrites API. To give more time to advanced users, BP plugin or theme developers so that they make their code ready for this API, unlike how we deal with deprecated code since r13304 & for the 2 following major releases (14.0 & 15.0), we will carry on loading it to prevent potential errors.

Props emaralive, espellcaste.

See #9210 (trunk)
Closes https://github.com/buddypress/buddypress/pull/352



git-svn-id: http://buddypress.svn.wordpress.org/trunk@13998 cdf35c40-ae34-48e0-9cc9-0c9da1808c22