MySQL: fix an error when creating the invitations API tables.
MySQL 8+ introduced a change where it no longer accepts a `string` as the default value for the `TEXT` data type. A literal default needs to be specified as expression instead.

To avoid breaking back-compat, we will default the data type to `NULL`, similar to how we do it in similar schema deta types in BuddyPress.

Fixes #9261
See https://github.com/buddypress/wp-cli-buddypress/pull/113

git-svn-id: http://buddypress.svn.wordpress.org/trunk@14100 cdf35c40-ae34-48e0-9cc9-0c9da1808c22