Avoid an error attempting to convert a `WP_Error` object to a `string` when checking for profile data existence.
While checking if a profile data exist, let us improve our error handling to avoid a PHP error trying to convert a `WP_Error` object to a `string`.

The reason why we might get a `WP_Error` object is varied, but it implies the data does not exist, and we should fail gracefully rather than throw an error.

Props rollybueno and potcus.

Closes https://github.com/buddypress/buddypress/pull/412/
Fixes #9290 (trunk)

git-svn-id: http://buddypress.svn.wordpress.org/trunk@14171 cdf35c40-ae34-48e0-9cc9-0c9da1808c22