When the `Extended Profiles` component is disabled, the "Profile Visibility" settings link does not 404.
The "Profile Visibility" link, `members/user/settings/profile/`, relies on the Extended Profiles component to work properly. When this component is disabled, the link should not added to the user profile settings menu.

We also did the same for the `Email` link. So that it is added only when the `Notifications` component is enabled.

Props joelkarunungan, nikunj8866, and vapvarun.

Closes https://github.com/buddypress/buddypress/pull/414/
Fixes #9297 (trunk)

git-svn-id: http://buddypress.svn.wordpress.org/trunk@14179 cdf35c40-ae34-48e0-9cc9-0c9da1808c22